network-intrusion-detection_with_QnA/network-intrusion-detection/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0
flask>=2.0.0
flask-socketio>=5.0.0
requests>=2.25.0
pathlib

# GPU LLM Dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0